+++ ProdManager/helpers/auth.py
@@ -6,8 +6,6 @@ from flask import session, redirect, url_for, g, abort
 def retreiv_auth():
   logged_until = session.get("logged_until", None)
 
-  #raise Exception(logged_until)
-
   if logged_until and (datetime.now() > datetime.fromtimestamp(logged_until)):
     session.clear()